+++ lloca/backbone/attention_backends/__init__.py
@@ -6,6 +6,7 @@ from importlib import metadata
 import torch
 
 # common kwargs used in custom attention backends
+VARLEN_KWARGS = ["cu_seq_q", "cu_seq_k", "max_q", "max_k"]
 XFORMERS_KWARGS = ["attn_bias", "op"]
 FLEX_KWARGS = ["score_mod", "block_mask"]
 FLASH_KWARGS = ["cu_seqlens_q", "cu_seqlens_k", "max_seqlen_q", "max_seqlen_k"]
@@ -31,25 +32,13 @@ for ep in metadata.entry_points(group="lloca.backbone.attention_backends"):
     _REGISTRY[ep.name] = module
 
 
-def get_sparse_attention_mask(*args, **kwargs):
-    """
-    Wrapper function to access xformers attention mask function if available.
-    """
-    try:
-        return _REGISTRY["xformers_attention"].get_xformers_attention_mask(*args, **kwargs)
-    except KeyError as err:
-        raise RuntimeError(
-            "Sparse attention masks currently only supported using xformers. "
-            "Run 'pip install lloca[xformers_attention]'."
-        ) from err
-
-
 def get_attention_backend(**kwargs):
     """
     Dynamically determine the attention backend based on the extra keyword arguments.
 
     Implemented backends:
     - PyTorch's native attention: torch.nn.functional.scaled_dot_product_attention
+    - PyTorch's varlen attention: torch.nn.attention.varlen.varlen_attn
     - Xformers attention: xformers.ops.memory_efficient_attention
     - PyTorch's flex_attention: torch.nn.attention.flex_attention.flex_attention
     - Original flash attention (supports variable sequence length): flash_attn.flash_attn_varlen_func
@@ -60,7 +49,9 @@ def get_attention_backend(**kwargs):
         return _REGISTRY[backend].attention
 
     # automatic fall-back based on other **kwargs
-    if any(kwargs.get(kwarg, None) is not None for kwarg in XFORMERS_KWARGS):
+    if any(kwargs.get(kwarg, None) is not None for kwarg in VARLEN_KWARGS):
+        return _REGISTRY["varlen"].attention
+    elif any(kwargs.get(kwarg, None) is not None for kwarg in XFORMERS_KWARGS):
         return _REGISTRY["xformers"].attention
     elif any(kwargs.get(kwarg, None) is not None for kwarg in FLEX_KWARGS):
         return _REGISTRY["flex"].attention